pyproject.toml
[project]
name = "dit-admin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "django==5.2",
    "django-cors-headers==4.7.0",
    "django-unfold>=0.60.0",
    "djangorestframework==3.16.0",
    "drf-yasg==1.21.10",
    "inflection==0.5.1",
    "packaging==24.2",
    "pillow>=11.3.0",
    "psycopg2==2.9.10",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "requests>=2.32.4",
    "sqlparse==0.5.3",
    "tzdata==2025.2",
    "uritemplate==4.1.1",
    "web3>=7.14.1",
]
